Implement Azure AD authentication handler and refactor service collection
