+++ TMS.App/MauiProgram.cs
@@ -8,6 +8,8 @@ using TMS.App.Pages;
 using TMS.App.ViewModels;
 using TMS.App.Controls;
 using Microsoft.Maui.Handlers;
+using TMS.Domain.Configuration;
+using TMS.Infrastructure.Extensions;
 
 
 namespace TMS.App
@@ -57,6 +59,9 @@ namespace TMS.App
 				}
 			#endif
 			});
+
+			// Options
+			builder.Services.Configure<AzureAdOptions>(builder.Configuration.GetRequiredSection("AzureAD"));
 			
 
 			// Other Services
