+++ TMS.App/TMS.App.csproj
@@ -93,6 +93,7 @@
 	<ItemGroup>
 	  <ProjectReference Include="..\TMS.Aspire.ServiceDefaults\TMS.Aspire.ServiceDefaults.csproj" />
 	  <ProjectReference Include="..\TMS.Application\TMS.Application.csproj" />
+	  <ProjectReference Include="..\TMS.Infrastructure\TMS.Infrastructure.csproj" />
 	</ItemGroup>
 
 </Project>
