+++ TMS.Domain/Configuration/AzureAdOptions.cs
@@ -1,0 +1,10 @@
+
+namespace TMS.Domain.Configuration
+{
+    public record AzureAdOptions
+    {
+        public required string RedirectUri { get; init; }
+        public required string ClientId { get; init; }
+        public required string[] Scopes { get; init; }
+    }
+}
