+++ TMS.Infrastructure/TMS.Infrastructure.csproj
@@ -26,7 +26,12 @@
 	</PropertyGroup>
 
 	<ItemGroup>
+		<PackageReference Include="Microsoft.Identity.Client" Version="4.81.0" />
 		<PackageReference Include="Microsoft.Maui.Controls" Version="$(MauiVersion)" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <ProjectReference Include="..\TMS.Domain\TMS.Domain.csproj" />
 	</ItemGroup>
 
 </Project>
